refactor: remove the repeated code to the from implementation in rust structure and some other minor review fixes
